requirements.txt
fastapi
uvicorn
websockets
langchain
langchain-community
langchain-qdrant
langchain-openai
langchain-text-splitters
qdrant-client
sentence-transformers
openai
datasets
python-dotenv
pydantic
PyYAML
sqlalchemy
psycopg[binary]
redis
PyJWT
cryptography
argon2-cffi
